+++ server/modules/reporting/api/index.js
@@ -1,0 +1,5 @@
+import * as reportingApi from './reportingApi'
+
+export const init = (app) => {
+    reportingApi.init(app)
+}
