+++ server/system/apiRouter.js
@@ -11,6 +11,7 @@ import * as collectImportApi from '@server/modules/collectImport/api/collectImpo
 import * as arenaImportApi from '@server/modules/arenaImport/api/arenaImportApi'
 import * as recordApi from '@server/modules/record/api/recordApi'
 import * as surveyRdbApi from '@server/modules/surveyRdb/api/surveyRdbApi'
+import * as reportingApi from '@server/modules/reporting/api'
 import * as analysisApi from '@server/modules/analysis/api'
 import * as jobApi from '@server/job/jobApi'
 import * as expressionApi from '@server/modules/expression/api/expressionApi'
@@ -35,6 +36,7 @@ arenaImportApi.init(router)
 recordApi.init(router)
 
 surveyRdbApi.init(router)
+reportingApi.init(router)
 
 analysisApi.init(router)
