--- webapp/components/DataQuery/Visualizer/Chart/index.js
@@ -1,1 +1,0 @@
-export { default } from './Chart'
