+++ webapp/components/DataQuery/Visualizer/Visualizer.js
@@ -4,11 +4,9 @@ import PropTypes from 'prop-types'
 import { Query } from '@common/model/query'
 
 import Table from './Table'
-import Chart from './Chart'
 
 const components = {
   [Query.displayTypes.table]: Table,
-  [Query.displayTypes.chart]: Chart,
 }
 
 const Visualizer = (props) => {
