+++ webapp/views/App/views/Data/Charts/Panel/index.js
@@ -1,0 +1,1 @@
+export { default } from './Panel'
